+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 - Fixed capitalization of URL property on PushSubscriptionRequestType.
 - Fixed capitalization of redirect element in autodiscovery.
+- Updated contact/update example to use a constant for the field URI.
 
 ## 1.0.0-rc.1 - 2017-03-02